+++ packages/shopify-app-remix/src/server/__tests__/shopify-app.test.ts
@@ -10,7 +10,6 @@ import {
   AppDistribution,
   ApiVersion,
 } from '../index';
-import {AppConfigArg} from '../config-types';
 import {testConfig} from '../__test-helpers';
 
 describe('shopifyApp', () => {
@@ -111,10 +110,10 @@ describe('shopifyApp', () => {
 
   it('fails if no session storage is given', () => {
     // GIVEN
-    const config: AppConfigArg = testConfig({sessionStorage: undefined});
+    const config = testConfig({sessionStorage: undefined});
     delete (config as any).sessionStorage;
 
     // THEN
-    expect(() => shopifyApp(config)).toThrowError(ShopifyError);
+    expect(() => shopifyApp(config as any)).toThrowError(ShopifyError);
   });
 });