+++ packages/shopify-app-remix/src/server/clients/admin/types.ts
@@ -1,4 +1,5 @@
 import {Session, ShopifyRestResources} from '@shopify/shopify-api';
+import {AdminOperations} from '@shopify/admin-api-client';
 
 import {BasicParams} from '../../types';
 import {GraphQLClient} from '../types';
@@ -29,7 +30,19 @@ export interface AdminApiContext<
    *
    * @example
    * <caption>Using REST resources.</caption>
-   * <description>Getting the number of orders in a store using REST resources.</description>
+   * <description>Getting the number of orders in a store using REST resources. Visit the [Admin REST API references](/docs/api/admin-rest) for examples on using each resource. </description>
+   *
+   * ```ts
+   * // /app/routes/**\/*.ts
+   * import { LoaderFunctionArgs, json } from "@remix-run/node";
+   * import { authenticate } from "../shopify.server";
+   *
+   * export const loader = async ({ request }: LoaderFunctionArgs) => {
+   *   const { admin, session } = await authenticate.admin(request);
+   *   return json(admin.rest.resources.Order.count({ session }));
+   * };
+   * ```
+   *
    *
    * ```ts
    * // /app/shopify.server.ts
@@ -44,6 +57,11 @@ export interface AdminApiContext<
    * export const authenticate = shopify.authenticate;
    * ```
    *
+   *
+   * @example
+   * <caption>Performing a GET request to the REST API.</caption>
+   * <description>Use `admin.rest.get` to make custom requests to make a request to to the `customer/count` endpoint</description>
+   *
    * ```ts
    * // /app/routes/**\/*.ts
    * import { LoaderFunctionArgs, json } from "@remix-run/node";
@@ -51,13 +69,11 @@ export interface AdminApiContext<
    *
    * export const loader = async ({ request }: LoaderFunctionArgs) => {
    *   const { admin, session } = await authenticate.admin(request);
-   *   return json(admin.rest.resources.Order.count({ session }));
+   *   const response = await admin.rest.get({ path: "/customers/count.json" });
+   *   const customers = await response.json();
+   *   return json({ customers });
    * };
    * ```
-   *
-   * @example
-   * <caption>Performing a GET request to the REST API.</caption>
-   * <description>Use `admin.rest.<method>` to make custom requests to the API.</description>
    *
    * ```ts
    * // /app/shopify.server.ts
@@ -71,7 +87,9 @@ export interface AdminApiContext<
    * export default shopify;
    * export const authenticate = shopify.authenticate;
    * ```
-   *
+   * @example
+   * <caption>Performing a POST request to the REST API.</caption>
+   * <description>Use `admin.rest.post` to make custom requests to make a request to to the `customers.json` endpoint to send a welcome email</description>
    * ```ts
    * // /app/routes/**\/*.ts
    * import { LoaderFunctionArgs, json } from "@remix-run/node";
@@ -79,10 +97,34 @@ export interface AdminApiContext<
    *
    * export const loader = async ({ request }: LoaderFunctionArgs) => {
    *   const { admin, session } = await authenticate.admin(request);
-   *   const response = await admin.rest.get({ path: "/customers/count.json" });
-   *   const customers = await response.json();
-   *   return json({ customers });
+   *   const response = admin.rest.post({
+   *     path: "customers/7392136888625/send_invite.json",
+   *     body: {
+   *       customer_invite: {
+   *         to: "new_test_email@shopify.com",
+   *         from: "j.limited@example.com",
+   *         bcc: ["j.limited@example.com"],
+   *         subject: "Welcome to my new shop",
+   *         custom_message: "My awesome new store",
+   *       },
+   *     },
+   * });
+   *   const customerInvite = await response.json();
+   *   return json({ customerInvite });
    * };
+   * ```
+   *
+   * ```ts
+   * // /app/shopify.server.ts
+   * import { shopifyApp } from "@shopify/shopify-app-remix/server";
+   * import { restResources } from "@shopify/shopify-api/rest/admin/2023-04";
+   *
+   * const shopify = shopifyApp({
+   *   restResources,
+   *   // ...etc
+   * });
+   * export default shopify;
+   * export const authenticate = shopify.authenticate;
    * ```
    */
   rest: RestClientWithResources<Resources>;
@@ -120,5 +162,5 @@ export interface AdminApiContext<
    * }
    * ```
    */
-  graphql: GraphQLClient;
+  graphql: GraphQLClient<AdminOperations>;
 }