+++ packages/shopify-app-remix/src/server/clients/storefront/types.ts
@@ -1,3 +1,5 @@
+import {StorefrontOperations} from '@shopify/storefront-api-client';
+
 import {GraphQLClient} from '../types';
 
 export interface StorefrontContext {
@@ -12,18 +14,18 @@ export interface StorefrontContext {
    * <caption>Querying the GraphQL API.</caption>
    * <description>Use `storefront.graphql` to make query / mutation requests.</description>
    * ```ts
-   * import { ActionFunctionArgs } from "@remix-run/node";
+   * // app/routes/**\/.ts
+   * import { json } from "@remix-run/node";
    * import { authenticate } from "../shopify.server";
    *
    * export async function action({ request }: ActionFunctionArgs) {
-   *   const { storefront } = await authenticate.storefront(request);
+   *   const { storefront } = await authenticate.public.appProxy(request);
    *
    *   const response = await storefront.graphql(`{blogs(first: 10) { edges { node { id } } } }`);
    *
-   *   const productData = await response.json();
-   *   return json({ data: productData.data });
+   *   return json(await response.json());
    * }
    * ```
    */
-  graphql: GraphQLClient;
+  graphql: GraphQLClient<StorefrontOperations>;
 }