+++ packages/shopify-app-remix/src/server/shopify-app.ts
@@ -21,7 +21,7 @@ import {
 } from './types';
 import {SHOPIFY_REMIX_LIBRARY_VERSION} from './version';
 import {registerWebhooksFactory} from './authenticate/webhooks';
-import {AuthStrategy} from './authenticate/admin/authenticate';
+import {authStrategyFactory} from './authenticate/admin/authenticate';
 import {authenticateWebhookFactory} from './authenticate/webhooks/authenticate';
 import {overrideLogger} from './override-logger';
 import {addDocumentResponseHeadersFactory} from './authenticate/helpers';
@@ -29,6 +29,7 @@ import {loginFactory} from './authenticate/login/login';
 import {unauthenticatedAdminContextFactory} from './unauthenticated/admin';
 import {authenticatePublicFactory} from './authenticate/public';
 import {unauthenticatedStorefrontContextFactory} from './unauthenticated/storefront';
+import {AuthCodeFlowStrategy} from './authenticate/admin/strategies/auth-code-flow';
 
 /**
  * Creates an object your app will use to interact with Shopify.
@@ -39,6 +40,7 @@ import {unauthenticatedStorefrontContextFactory} from './unauthenticated/storefr
  * @example
  * <caption>The minimum viable configuration</caption>
  * ```ts
+ * // /shopify.server.ts
  * import { shopifyApp } from "@shopify/shopify-app-remix/server";
  *
  * const shopify = shopifyApp({
@@ -64,7 +66,11 @@ export function shopifyApp<
   }
 
   const params: BasicParams = {api, config, logger};
-  const oauth = new AuthStrategy<Config, Resources>(params);
+  const oauth = new AuthCodeFlowStrategy(params);
+  const authStrategy = authStrategyFactory<Config, Resources>({
+    ...params,
+    strategy: oauth,
+  });
 
   const shopify:
     | AdminApp<Config>
@@ -74,7 +80,7 @@ export function shopifyApp<
     addDocumentResponseHeaders: addDocumentResponseHeadersFactory(params),
     registerWebhooks: registerWebhooksFactory(params),
     authenticate: {
-      admin: oauth.authenticateAdmin.bind(oauth),
+      admin: authStrategy,
       public: authenticatePublicFactory<Config['future'], Resources>(params),
       webhook: authenticateWebhookFactory<
         Config['future'],
@@ -142,6 +148,7 @@ function deriveApi(appConfig: AppConfigArg) {
     isEmbeddedApp: appConfig.isEmbeddedApp ?? true,
     apiVersion: appConfig.apiVersion ?? LATEST_API_VERSION,
     isCustomStoreApp: appConfig.distribution === AppDistribution.ShopifyAdmin,
+    future: {},
   });
 }