+++ packages/shopify-app-express/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## 4.0.0
+
+### Major Changes
+
+- d3e4b5e: Updated `@shopify/shopify-api` to the latest major version. Please follow [the v9 migration guide](https://github.com/Shopify/shopify-api-js/blob/main/packages/shopify-api/docs/migrating-to-v9.md) to update your app.
+
+### Patch Changes
+
+- Updated dependencies [d3e4b5e]
+  - @shopify/shopify-app-session-storage-memory@2.0.3
+  - @shopify/shopify-app-session-storage@2.0.3
+
+## 3.0.2
+
+### Patch Changes
+
+- 3685bd4: Bump shopify-api to ^8.1.1
+- Updated dependencies [3685bd4]
+  - @shopify/shopify-app-session-storage-memory@2.0.2
+  - @shopify/shopify-app-session-storage@2.0.2
+
 ## 3.0.1
 
 ### Patch Changes